refactor: move the shared search-box helpers into _initialization_box
